pyproject.toml
[project]
name = "pallas-agent"
version = "0.1.0"
description = "Pallas: A self-evolving autonomous AI agent built by Vinkura AI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Vinkura AI" }]
keywords = ["ai", "agent", "llm", "automation", "cli"]

dependencies = [
    "anthropic>=0.50.0",
    "google-genai>=1.0.0",
    "openai>=2.0.0",
    "pydantic>=2.10.0",
    "rich>=14.0.0",
    "httpx>=0.28.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.5.0"]
telegram = ["pyTelegramBotAPI>=4.18.0"]
discord = ["discord.py>=2.3.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
pallas = "pallas_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "pallas_adapter",
    "pallas_registry",
    "pallas_core",
    "pallas_cli",
    "gateway",
    "cron",
    "environments",
    "tools",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
